Update 32-bit to 64-bit system
- Fix data type inconsistency issue
	- uint32_t and index_t misuse in original code

+++ include/groute/graphs/common.h
@@ -33,7 +33,7 @@
 #include <stdint.h>
 #include <cuda_runtime.h>
 
-typedef uint32_t index_t;
+typedef uint64_t index_t;
 
 
 #endif // __GROUTE_GRAPHS_COMMON_H
